1.공지사항 완료 2.패키지 변경
