+++ workspace/semiPrj/src/coty/admin/notice/kyw/NoticeView_c_Controller.java
@@ -1,4 +1,4 @@
-package coty.member.notice;
+package coty.admin.notice.kyw;
 
 import java.io.IOException;
 
@@ -8,7 +8,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import coty.admin.noticeVo.Notice_a_Vo;
+import coty.admin.notice.kyw.Notice_a_Vo;
 import coty.admin.service.NoticeService;
