+++ workspace/semiPrj/src/coty/admin/notice/kyw/Notice_a_Vo.java
@@ -1,4 +1,4 @@
-package coty.admin.noticeVo;
+package coty.admin.notice.kyw;
 
 public class Notice_a_Vo {
