+++ workspace/semiPrj/src/coty/admin/notice/kyw/Notice_deleteController.java
@@ -1,4 +1,4 @@
-package coty.admin.controller;
+package coty.admin.notice.kyw;
 
 import java.io.IOException;
 
@@ -7,8 +7,9 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
+import coty.admin.notice.kyw.Notice_a_Vo;
 import coty.admin.service.NoticeService;
+
 
 
 @WebServlet("/admin/Notice_delete")
