+++ workspace/semiPrj/src/coty/admin/notice/kyw/Notice_editController.java
@@ -1,4 +1,4 @@
-package coty.admin.controller;
+package coty.admin.notice.kyw;
 
 import java.io.IOException;
 
@@ -8,10 +8,10 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+import coty.admin.notice.kyw.Notice_a_Vo;
 
-import coty.admin.noticeVo.Notice_a_Vo;
 import coty.admin.service.NoticeService;
-import coty.util.PageVo;
 
 @MultipartConfig(
 		maxFileSize = 1024 * 1024 * 50,
@@ -62,43 +62,41 @@ public class Notice_editController extends HttpServlet {
 	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		
 		//데이터 꺼내기 //제목,내용,번호
+		HttpSession s = req.getSession();
+		Notice_a_Vo Notice_a_Vo = (Notice_a_Vo)s.getAttribute("Notice_a_Vo");
+		
 		String Title = req.getParameter("Title");
 		String writer = req.getParameter("writer");
 		String Content = req.getParameter("Content");
-		
 		//데이터 뭉치기
 		Notice_a_Vo vo = new Notice_a_Vo();
+		vo.setNo(Notice_a_Vo.getNo());
 		vo.setTitle(Title);
-		vo.setNo(writer);
+		vo.setWriter(writer);
 		vo.setContent(Content);
+
 		
 		//서비스 로직	//클라한테 전달받은 데이터들을 가지고 업데이트 쿼리 실행
 		int result = 0;
 		try {
 			NoticeService ns = new NoticeService();	
-			//result = ns.update(writer, Title, Content);	//xxx
-			int vo = ns.update(writer, Title, Content);
+			result = ns.update(vo);	
 			
-			req.setAttribute("Notice_a_Vo", vo);
-			req.getRequestDispatcher("/WEB-INF/views/admin/Notice_edit.jsp").forward(req, resp);
-			System.out.println(vo);
 		} catch (Exception e) {
 			System.out.println("에러 ㅠㅠ");
 			e.printStackTrace();
 		}
 		
+
+		if(result == 1) {
+		req.getSession().setAttribute("alertMsg", "게시글 수정 성공");
+		resp.sendRedirect("/admin/Notice_list?page=1");
+		}else{
+			req.getSession().setAttribute("alertMsg", "게시글 수정 실패..");
+			resp.sendRedirect("/");	
+		}
 		
-				
-				//서비스레이어.업데이트쿼리실행하느넴소드(뭉친데이터);
-		
-		//화면
-//		if() {
-//			
-//		}else{
-//			
-//		}
-		
-		
+		s.invalidate();
 	}
