+++ workspace/semiPrj/src/coty/admin/notice/kyw/Notice_writeController.java
@@ -1,4 +1,4 @@
-package coty.admin.controller;
+package coty.admin.notice.kyw;
 
 import java.io.IOException;
 
@@ -9,7 +9,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import coty.admin.noticeVo.Notice_a_Vo;
+import coty.admin.notice.kyw.Notice_a_Vo;
 import coty.admin.service.NoticeService;
 
 @MultipartConfig(
